Extract jsonSchemaToTs renderer into its own module + spec
Moves the JSON Schema → TypeScript renderer (jsonSchemaToTs plus all
its helpers: schema merging, ref resolution, if/then/else narrowing,
literal rendering, etc.) and its full test suite out of the MCP-tool
module into dedicated files:

- src/.../tools/json_schema_to_ts.ts       (renderer + 12-item
                                            deviation docstring)
- src/.../tools/json_schema_to_ts.spec.ts  (60 tests mirroring
                                            json-schema-to-ts coverage)

mcp_type_defs.ts now imports jsonSchemaToTs from the new module and
keeps only the MCP-specific concerns: McpToolDef, collectMcpToolDefs,
buildMcpCapabilityMap, buildMcpTypeDefsBlock.

No functional change. All tests still pass.
